add monarq sim to integration tests
